Add copyright notice and update project metadata
Added a copyright notice for the year 2024 in the `Directory.Build.props` file.
Updated `Infinity.Toolkit.FeatureModules.csproj` with a description and tags to clarify its purpose in simplifying development in feature slices.
Updated `Infinity.Toolkit.LogFormatter.csproj` with a description and tags to highlight its logging formatter capabilities with a Visual Studio Code inspired theme.